partition: Avoid overflow.
 Error: OVERFLOW_BEFORE_WIDEN (CWE-190): [#def10]
 nbdkit-1.15.7/filters/partition/partition-mbr.c:89: overflow_before_widen: Potentially overflowing expression "partition.start_sector * 512U" with type "unsigned int" (32 bits, unsigned) is evaluated using 32-bit arithmetic, and then used in a context that expects an expression of type "int64_t" (64 bits, signed).
nbdkit-1.15.7/filters/partition/partition-mbr.c:89: remediation: To avoid overflow, cast either "partition.start_sector" or "512U" to type "int64_t".
 #   87|             !is_extended (partition.part_type_byte) &&
 #   88|             partnum == i+1) {
 #   89|->         *offset_r = partition.start_sector * SECTOR_SIZE;
 #   90|           *range_r = partition.nr_sectors * SECTOR_SIZE;
 #   91|           return 0;

 Error: OVERFLOW_BEFORE_WIDEN (CWE-190): [#def11]
 nbdkit-1.15.7/filters/partition/partition-mbr.c:90: overflow_before_widen: Potentially overflowing expression "partition.nr_sectors * 512U" with type "unsigned int" (32 bits, unsigned) is evaluated using 32-bit arithmetic, and then used in a context that expects an expression of type "int64_t" (64 bits, signed).
nbdkit-1.15.7/filters/partition/partition-mbr.c:90: remediation: To avoid overflow, cast either "partition.nr_sectors" or "512U" to type "int64_t".
 #   88|             partnum == i+1) {
 #   89|           *offset_r = partition.start_sector * SECTOR_SIZE;
 #   90|->         *range_r = partition.nr_sectors * SECTOR_SIZE;
 #   91|           return 0;
 #   92|         }

Bug found by Coverity.

+++ filters/partition/partition-mbr.c
@@ -86,8 +86,8 @@ find_mbr_partition (struct nbdkit_next_ops *next_ops, void *nxdata,
           partition.part_type_byte != 0 &&
           !is_extended (partition.part_type_byte) &&
           partnum == i+1) {
-        *offset_r = partition.start_sector * SECTOR_SIZE;
-        *range_r = partition.nr_sectors * SECTOR_SIZE;
+        *offset_r = partition.start_sector * (int64_t) SECTOR_SIZE;
+        *range_r = partition.nr_sectors * (int64_t) SECTOR_SIZE;
         return 0;
       }
     }
